+++ src/useTimesplitter/index.js
@@ -5,98 +5,9 @@ import {
   useEffect,
   useMemo,
 } from "react";
+import reducer from "./reducer";
+import createActions from "./actions";
 import { totalTime } from "../lib";
-
-async function refreshContent(url) {
-  try {
-    return await fetch(url).then((res) => res.json());
-  } catch (error) {
-    console.error(`An error occurred fetching content from ${url}`);
-    throw error;
-  }
-}
-
-function course(state, action = {}) {
-  switch (action.type) {
-    case "LOADED":
-      return action.payload;
-    default:
-      return state;
-  }
-}
-
-function loading(state = false, action = {}) {
-  switch (action.type) {
-    case "LOAD":
-      return true;
-    case "LOADED":
-      return false;
-    case "ERROR":
-      return false;
-    default:
-      return state;
-  }
-}
-
-function preview(state = true, action = {}) {
-  switch (action.type) {
-    case "START":
-      return false;
-    default:
-      return state;
-  }
-}
-
-const samples = [
-  { title: "this topic", time: { est: 5 } },
-  { title: "that topic", time: { est: 2 } },
-  { title: "the other topic", time: { est: 12 } },
-  { title: "the main topic", time: { est: 3 } },
-];
-
-function adjust(
-  state = {
-    manualAdjust: false,
-    minsOver: 10,
-    endTime: new Date(),
-    remainingTopics: samples,
-  },
-  action = {}
-) {
-  switch (action.type) {
-    default:
-      return state;
-  }
-}
-
-function reducer(state = null, action = {}) {
-  return {
-    loading: loading(state.loading, action),
-    preview: preview(state.preview, action),
-    adjust: adjust(state.adjust, action),
-    course: course(state.course, action),
-  };
-}
-
-function createActions(dispatch = (f) => f, contentUrl) {
-  return {
-    refresh() {
-      dispatch({ type: "LOAD" });
-      refreshContent(contentUrl).then((payload) => {
-        dispatch({ type: "LOADED", payload });
-      });
-    },
-    start() {
-      dispatch({ type: "START" });
-    },
-    end() {
-      dispatch({ type: "END" });
-    },
-    adjust(length) {
-      dispatch({ type: "ADJUST", payload: { length } });
-    },
-  };
-}
 
 const TimesplitterContext = createContext();
 export const useTimesplitter = () => useContext(TimesplitterContext);
