+++ src/useTimesplitter/reducer/course.js
@@ -1,0 +1,8 @@
+export default function course(state, action = {}) {
+  switch (action.type) {
+    case "LOADED":
+      return action.payload;
+    default:
+      return state;
+  }
+}
