+++ src/useTimesplitter/reducer/preview.js
@@ -1,0 +1,8 @@
+export default function preview(state = true, action = {}) {
+  switch (action.type) {
+    case "START":
+      return false;
+    default:
+      return state;
+  }
+}
